+++ l1/src/messages/wheelsendmessage.h
@@ -5,6 +5,15 @@
 #include <boost/property_tree/ptree.hpp>
 
 
+/**
+ * The WheelSendMessage encapsulates all commands that can be sent to a client.
+ *
+ * As a result, any WheelSendMessage will generate the following JSON string:
+ * {"id":{"PWM":int,"ROT":int,"KP":double,"KI":double,"KD":double}}
+ *
+ * @brief The WheelSendMessage class encapsulates all commands that can be sent
+ * to the microcontroller
+ */
 class WheelSendMessage
 {
     friend class Builder;
@@ -16,17 +25,28 @@ public:
     static constexpr const char *KEY_DIFFERENTIAL = "KD";
 
 
+    /**
+     * The WheelSendMessage::Builder generates a message
+     * in a convienient way.
+     * @brief The Builder class build WheelSendMessage
+     */
     class Builder
     {
         friend class WheelSendMessage;
         int id {0};
         int pwm {0};
-        int angularVelocity {0};
+        double angularVelocity {0};
         double kp {0.0};
         double ki {0.0};
         double kd {0.0};
 
     public:
+        /**
+         * @brief Builder Creates a Builder to generate
+         * a WheelSendMessage instance
+         * @param id an ID of a wheel. The ID is a mean
+         * to choose a device where a command can be issue.
+         */
         explicit Builder(int id);
 
         Builder(const Builder &) = delete;
@@ -34,15 +54,61 @@ public:
         Builder &operator=(const Builder &) = delete;
         Builder &operator=(Builder &&) = delete;
 
+        /**
+         * @brief setPwm Sets PWM in a wheel of the given ID
+         * @param pwm PWM value. The values must be consulted with
+         * the documentation of a wheel driver provider.
+         * @return  Builder reference
+         */
         Builder &setPwm(int pwm) noexcept;
-        Builder &setAngularVelocity(int angularVelocity) noexcept;
+
+        /**
+         * @brief setAngularVelocity Sets angular velocity in a wheel
+         * of the given ID
+         * @param angularVelocity Angular velocity
+         * @return  Builder reference
+         */
+        Builder &setAngularVelocity(double angularVelocity) noexcept;
+
+        /**
+         * @brief setProportional Sets proportional constant
+         * and allows sending it to the client device.
+         * @param kp proportional coefficient
+         * @return  Builder reference
+         */
         Builder &setProportional(double kp) noexcept;
+
+        /**
+         * @brief setIntegral Sets integral constant
+         * and allows sending it to the client device.
+         * @param ki integral coefficient
+         * @return  Builder reference
+         */
         Builder &setIntegral(double ki) noexcept;
+
+        /**
+         * @brief setDifferential Sets differential constant
+         * and allows sending it to the client device.
+         * @param kd differential coefficient
+         * @return  Builder reference
+         */
         Builder &setDifferential(double kd) noexcept;
+
+        /**
+         * A builder executor. It build a final object.
+         * @brief build Builds WheelSendMessage
+         * @return WheelSendMessage
+         */
         WheelSendMessage build();
     };
 
 
+    /**
+     * @brief toString Generates string representation of JSON
+     * values
+     * @param pretty formats the data so it looks pretty
+     * @return string representation of JSON
+     */
     std::string toString(bool pretty = false) const;
 
 private:
